third_party/blink/web_tests/http/tests/websocket/send-blob-onmessage-origin-expected.txt

WebSocket message event origin attribute test (Blob message)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: Message #0.
PASS origin is "ws://127.0.0.1:8880"
PASS PASS: Message #1.
PASS origin is "ws://127.0.0.1:8880"
PASS PASS: Message #2.
PASS origin is "ws://127.0.0.1:8880"
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
